Refactor: Move theme logic into LangService
The ThemeService is no longer needed and its functionality has been merged into LangService.
