+++ plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/services/LangService.kt
@@ -2,6 +2,7 @@ package dev.betrix.superSmashMobsBrawl.services
 
 import java.io.InputStreamReader
 import java.nio.charset.StandardCharsets
+import java.util.Locale
 import java.util.regex.Matcher
 import java.util.regex.Pattern
 import net.kyori.adventure.text.Component
@@ -13,10 +14,21 @@ import org.koin.core.component.inject
 
 class LangService : KoinComponent {
     private val plugin: JavaPlugin by inject()
-    private val theme: ThemeService by inject()
 
     private val dataFolder = plugin.dataFolder
     private val enLang: YamlConfiguration = loadLangWithDefaults("data/lang/en.yml")
+    private val themeConfig: YamlConfiguration = loadLangWithDefaults("data/theme.yml")
+    private val tokenToHex: Map<String, String> by lazy {
+        val section = themeConfig.getConfigurationSection("tokens")
+        val map = linkedMapOf<String, String>()
+        if (section != null) {
+            for (key in section.getKeys(false)) {
+                val hex = section.getString(key) ?: continue
+                map[key.lowercase(Locale.ROOT)] = hex
+            }
+        }
+        map.toMap()
+    }
 
     private val miniMessage: MiniMessage = MiniMessage.miniMessage()
 
@@ -26,8 +38,8 @@ class LangService : KoinComponent {
     fun t(key: String, varsBuilder: VarsBuilder.() -> Unit = {}): Component {
         val vars = VarsBuilder().apply(varsBuilder).build()
         val resolved = resolveAndFormat(key, vars, visited = mutableSetOf())
-        val preprocessed = theme.preprocessMessageColors(resolved)
-        return miniMessage.deserialize(preprocessed, theme.tokenTagResolver())
+        val preprocessed = preprocessThemeTokens(resolved)
+        return miniMessage.deserialize(preprocessed)
     }
 
     private fun resolveAndFormat(
@@ -110,6 +122,61 @@ class LangService : KoinComponent {
 
         matcher.appendTail(sb)
         return sb.toString()
+    }
+
+    /**
+     * Replace token tags like <primary> and </primary> with concrete color tags
+     * using hex values from data/theme.yml. Also expand token names in arguments
+     * for tags that accept color parameters (e.g., <gradient:primary:ui-shadow>, <shadow:text>).
+     */
+    private fun preprocessThemeTokens(message: String): String {
+        if (tokenToHex.isEmpty()) return message
+
+        var result = message
+
+        // Replace tokens inside color-accepting tag arguments (gradient, shadow)
+        run {
+            val pattern = Pattern.compile("<(gradient|shadow):([^>]+)>")
+            val matcher = pattern.matcher(result)
+            val sb = StringBuffer()
+            while (matcher.find()) {
+                val tagName = matcher.group(1)
+                val args = matcher.group(2)
+                val replacedArgs = args.split(":").joinToString(":") { part ->
+                    val key = part.trim().lowercase(Locale.ROOT)
+                    tokenToHex[key] ?: part
+                }
+                val replacement = "<" + tagName + ":" + replacedArgs + ">"
+                matcher.appendReplacement(sb, Matcher.quoteReplacement(replacement))
+            }
+            matcher.appendTail(sb)
+            result = sb.toString()
+        }
+
+        // Replace simple token tags <token> and </token> with <hex> and </hex>
+        run {
+            val group = tokenToHex.keys.joinToString("|") { Pattern.quote(it) }
+            if (group.isNotEmpty()) {
+                val pattern = Pattern.compile("<(/?)($group)>")
+                val matcher = pattern.matcher(result)
+                val sb = StringBuffer()
+                while (matcher.find()) {
+                    val slash = matcher.group(1)
+                    val name = matcher.group(2).lowercase(Locale.ROOT)
+                    val hex = tokenToHex[name]
+                    if (hex != null) {
+                        val replacement = if (slash.isEmpty()) "<$hex>" else "</$hex>"
+                        matcher.appendReplacement(sb, Matcher.quoteReplacement(replacement))
+                    } else {
+                        matcher.appendReplacement(sb, Matcher.quoteReplacement(matcher.group()))
+                    }
+                }
+                matcher.appendTail(sb)
+                result = sb.toString()
+            }
+        }
+
+        return result
     }
 
     class VarsBuilder {
